+++ blocklets/ai-studio/src/components/trace/index.tsx
@@ -1,7 +1,8 @@
 import { ImagePreviewB64 } from '@app/pages/project/debug-view';
 import { ImageType, MessageInput, SessionItem } from '@app/pages/project/state';
 import { useLocaleContext } from '@arcblock/ux/lib/Locale/context';
-import { PromptMessages, RunAssistantInput, RunAssistantLog, RunAssistantUsage } from '@blocklet/ai-runtime/types';
+import { ChatCompletionInput } from '@blocklet/ai-kit/api/types/chat';
+import { RunAssistantInput, RunAssistantLog, RunAssistantUsage } from '@blocklet/ai-runtime/types';
 import { Accordion, AccordionDetails, AccordionSummary, Box, Stack, Typography, styled } from '@mui/material';
 import { GridExpandMoreIcon } from '@mui/x-data-grid';
 import dayjs from 'dayjs';
@@ -14,7 +15,7 @@ type Trace =
   | { label: 'startTime' | 'endTime'; value: number }
   | { label: 'logs'; value: RunAssistantLog[] }
   | { label: 'inputParameters'; value: { [key: string]: string } }
-  | { label: 'promptMessages'; value: PromptMessages }
+  | { label: 'promptMessages'; value: ChatCompletionInput['messages'] }
   | { label: 'output'; value: string }
   | { label: 'images'; value: ImageType }
   | { label: 'apiArgs'; value: any }