+++ blocklets/ai-studio/src/libs/deployment.ts
@@ -2,6 +2,7 @@ import { ProjectSettings } from '@blocklet/ai-runtime/types';
 import { joinURL } from 'ufo';
 
 import axios from './api';
+import { User } from './project';
 
 export type Deployment = {
   id: string;
@@ -17,14 +18,18 @@ export type Deployment = {
   categories: { id: string; name: string; slug: string }[];
   productHuntUrl?: string;
   productHuntBannerUrl?: string;
+  aigneBannerVisible?: boolean;
 };
 
+export type ProjectStatsItem = { projectId: string; totalRuns: number; totalUsers: number };
+
 export type UpdateType = {
-  access: 'private' | 'public';
+  access?: 'private' | 'public';
   categories?: string[];
   productHuntUrl?: string;
   productHuntBannerUrl?: string;
   orderIndex?: number;
+  aigneBannerVisible?: boolean;
 };
 
 export async function getDeploymentByProjectId({
@@ -58,7 +63,7 @@ export async function getDeploymentsByCategorySlug(input: {
   page: number;
   pageSize: number;
 }): Promise<{
-  list: (Deployment & { project: ProjectSettings })[];
+  list: (Deployment & { project: ProjectSettings; stats: ProjectStatsItem; createdByInfo: User })[];
   totalCount: number;
 }> {
   const { categorySlug, page, pageSize } = input;
@@ -75,11 +80,11 @@ export async function getDeployments(input: { page: number; pageSize: number }):
 }
 
 export async function updateDeployment(id: string, input: UpdateType): Promise<Deployment> {
-  return axios.put(joinURL('/api/deployments', id), input).then((res) => res.data);
+  return axios.patch(joinURL('/api/deployments', id), input).then((res) => res.data);
 }
 
 export async function adminUpdateDeployment(id: string, input: UpdateType): Promise<Deployment> {
-  return axios.put(joinURL('/api/admin/deployments', id), input).then((res) => res.data);
+  return axios.patch(joinURL('/api/admin/deployments', id), input).then((res) => res.data);
 }
 
 export async function deleteDeployment(input: { id: string }): Promise<Deployment> {