+++ blocklets/ai-studio/CHANGELOG.md
@@ -1,3 +1,88 @@
+## 0.4.119 (2024-10-31)
+
+- fix: invalid sig error
+
+## 0.4.118 (2024-10-30)
+
+- chore: upgrade @blocklet/sdk & custom notification for issued vc
+
+## 0.4.117 (2024-10-30)
+
+- chore: update @blocklet/uploader package
+
+## 0.4.116 (2024-10-30)
+
+- fix: auto prefix VCs with AIGNE for those issued by agent
+
+## 0.4.115 (2024-10-30)
+
+- chore: remove user passport caching
+- chore(api): add stats/userinfo to recommend project list response
+
+## 0.4.114 (2024-10-29)
+
+- fix: should remove agentId condition from memory query
+
+## 0.4.113 (2024-10-29)
+
+- fix: not found child agent in decision agent
+- fix: invalid input parameter in call agent
+
+## 0.4.112 (2024-10-29)
+
+- chore: set current tab when saved memory data
+- chore: close dialog when path changed
+- fix: unexpected memory behavior
+
+## 0.4.111 (2024-10-29)
+
+- fix: change knowledge createBy when copy knowledge
+
+## 0.4.110 (2024-10-29)
+
+- chore: polish explore list page
+
+## 0.4.109 (2024-10-28)
+
+- chore: improve UI for explore list page
+
+## 0.4.108 (2024-10-27)
+
+- refactor: optimize decision function name generator
+
+## 0.4.107 (2024-10-27)
+
+- fix: auto login when running agent
+
+## 0.4.106 (2024-10-27)
+
+- feat: support component schema filter & preview
+
+## 0.4.105 (2024-10-26)
+
+- fix: all fields must be required for json schema mode
+
+## 0.4.104 (2024-10-25)
+
+- feat: add vc input support
+
+## 0.4.103 (2024-10-25)
+
+- fix: support integer parameter for open api
+
+## 0.4.102 (2024-10-25)
+
+- fix: full height view for chat layout
+- chore: persistent aigne banner visibility & ui improvements
+
+## 0.4.101 (2024-10-25)
+
+- fix: auto open login dialog when running an agent
+
+## 0.4.100 (2024-10-24)
+
+- fix: parsing pdf content is deranged content
+
 ## 0.4.99 (2024-10-24)
 
 - feat: add preference to toggle premium plan
@@ -214,7 +299,7 @@
 
 ## 0.4.48 (2024-9-12)
 
-- fix: use quickjs instead of vm2 for better security
+- fix: use QuickJS instead of vm2 for better security
 
 ## 0.4.47 (2024-9-12)
 
@@ -472,7 +557,7 @@
 
 ## 0.3.27 (2024-7-11)
 
-- fix: multiple scrollbars of upload logo dialog
+- fix: multiple scrollbar of upload logo dialog
 
 ## 0.3.26 (2024-7-11)